// src/openapi/api/models/CommitteeWithSeats.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Diplomatiq API
 * This is the OpenAPI documentation of the Diplomatiq API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * The list of the committees of the conference with seats
 * @export
 * @interface CommitteeWithSeats
 */
export interface CommitteeWithSeats {
    /**
     * The name of the committee
     * @type {string}
     * @memberof CommitteeWithSeats
     */
    name: string;
    /**
     * The code name of the committee
     * @type {string}
     * @memberof CommitteeWithSeats
     */
    codeName: string;
    /**
     * The list of the represented countries in the committee as ISO3166-1 alpha-2 two letter country codes
     * @type {Array<string>}
     * @memberof CommitteeWithSeats
     */
    representedCountries: Array<string>;
}

export function CommitteeWithSeatsFromJSON(json: any): CommitteeWithSeats {
    return CommitteeWithSeatsFromJSONTyped(json, false);
}

export function CommitteeWithSeatsFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommitteeWithSeats {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'codeName': json['codeName'],
        'representedCountries': json['representedCountries'],
    };
}

export function CommitteeWithSeatsToJSON(value?: CommitteeWithSeats | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'codeName': value.codeName,
        'representedCountries': value.representedCountries,
    };
}